feat: Auth필요한 API 수정

+++ controller/CartController.js
@@ -72,10 +72,22 @@ const getCartItems = (req, res) => {
 
 // 장바구니 아이템 삭제 
 const removeCartItem = (req, res) => {
-    const cartItemId = req.params.id;
+    let authorization = ensureAuthorization(req, res);
 
-    let sql = "DELETE FROM cartItems WHERE id = ?;";
-    conn.query(sql, cartItemId, 
+    if(authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        });
+    } else if(authorization instanceof jwt.JsonWebTokenError) {
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    }
+    else {
+        const cartItemId = req.params.id;
+
+        let sql = "DELETE FROM cartItems WHERE id = ?;";
+        conn.query(sql, cartItemId, 
         (err, results) => {
             if(err) {
                 console.log(err);
@@ -83,7 +95,10 @@ const removeCartItem = (req, res) => {
             }
 
             return res.status(StatusCodes.OK).json(results);
-    })
+        })
+    }
+
+
 }
